+++ eye_tracking_mouse/utils/CoordinateSmoother.cs
